chore: commented, removed console logging, moved constants to separate file. #1
